Fixes made in sahil-name-sorter
