+++ Employee/sahil-name-sorter.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.28010.2003
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Employee", "Employee\Employee.csproj", "{A91301AE-A4E3-4485-B5A4-3DD75F369117}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "sahil-name-sorter", "Employee\sahil-name-sorter.csproj", "{A91301AE-A4E3-4485-B5A4-3DD75F369117}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
