--- Employee/sorted-names-list.txt
@@ -1,8 +1,0 @@
-Anthony, , B, David
-Archer, Post, Arthur, Buth
-Dean, , , James
-Doe, , Jimmy, John
-Farnsworth, , Fullhouse, Pribus
-Hammersworth, , X, Evelyn
-Luxley, , , Adalous
-O'Connoll, , , Richard
