--- Employee/unsorted-names-list.txt
@@ -1,8 +1,0 @@
-John,Jimmy,,Doe
-Adalous,,,Luxley
-Buth,Arthur,Post,Archer
-Evelyn,X,,Hammersworth
-David,B,,Anthony
-Pribus,Fullhouse,,Farnsworth
-James,,,Dean
-Richard,,,O'Connoll
